+++ obc/drivers/cc1120/cc1120_spi.h
@@ -3,6 +3,40 @@
 #include <stdint.h>
 #include "obc_errors.h"
 
+// See chapter 8.5 in the datasheet
+#define TXRX_INTERRUPT_THRESHOLD 100U
+
+typedef struct {
+  uint8_t addr;
+  uint8_t val;
+} register_setting_t;
+
+typedef enum {
+  CC1120_STATE_SLEEP = 0,
+  CC1120_STATE_IDLE,
+  CC1120_STATE_XOFF,
+  CC1120_STATE_BIAS_SETTLE_MC,
+  CC1120_STATE_REG_SETTLE_MC,
+  CC1120_STATE_MANCAL,
+  CC1120_STATE_BIAS_SETTLE,
+  CC1120_STATE_REG_SETTLE,
+  CC1120_STATE_STARTCAL,
+  CC1120_STATE_BWBOOST,
+  CC1120_STATE_FS_LOCK,
+  CC1120_STATE_IFADCON,
+  CC1120_STATE_RX,
+  CC1120_STATE_RX_END,
+  CC1120_STATE_RESERVED,
+  CC1120_STATE_TXRX_SWITCH,
+  CC1120_STATE_RX_FIFO_ERR,
+  CC1120_STATE_FSTXON,
+  CC1120_STATE_TX,
+  CC1120_STATE_TX_END,
+  CC1120_STATE_RXTX_SWITCH,
+  CC1120_STATE_TX_FIFO_ERR,
+  CC1120_STATE_IFADCON_TXRX
+} cc1120_state_t;
+
 /**
  * @brief - Reads from consecutive registers from the CC1120.
  *
@@ -106,3 +140,34 @@ obc_error_code_t cc1120WriteFifoDirect(uint8_t addr, uint8_t data[], uint8_t len
  * @return OBC_ERR_CODE_CC1120_INVALID_STATUS_BYTE - If the status byte is invalid.
  */
 obc_error_code_t cc1120SendByteReceiveStatus(uint8_t data);
+
+/**
+ * @brief Gets the number of bytes queued in the TX FIFO
+ *
+ * @param numBytes - A pointer to an 8-bit integer to store the number of bytes in
+ * @return obc_error_code_t - Whether or not the registe read was successful
+ */
+obc_error_code_t cc1120GetBytesInTxFifo(uint8_t *numBytes);
+
+/**
+ * @brief Gets the number of bytes queued in the RX FIFO
+ *
+ * @param numBytes - A pointer to an 8-bit integer to store the number of bytes in
+ * @return obc_error_code_t - Whether or not the register read was successful
+ */
+obc_error_code_t cc1120GetBytesInRxFifo(uint8_t *numBytes);
+
+/**
+ * @brief Gets the state of the CC1120 from the MARCSTATE register
+ *
+ * @param stateNum - A pointer to an 8-bit integer to store the state in
+ * @return obc_error_code_t - Whether or not the register read was successful
+ */
+obc_error_code_t cc1120GetState(cc1120_state_t *stateNum);
+
+/**
+ * @brief Resets CC1120 & initializes transmit mode
+ *
+ * @return obc_error_code_t - Whether or not the setup was a success
+ */
+obc_error_code_t cc1120Init(void);
