+++ obc/modules/comms_link_mgr/cc1120_txrx.c
@@ -6,7 +6,6 @@
 #include "obc_math.h"
 #include "obc_board_config.h"
 
-// TODO: Ideally, we shouldn't need to include this in the driver
 #include "uplink_decoder.h"
 
 #include <FreeRTOS.h>
@@ -39,57 +38,7 @@ static obc_error_code_t cc1120SendInifinitePktMode(uint8_t *data, uint32_t len);
 
 static obc_error_code_t writeFifoBlocking(uint8_t *data, uint32_t len);
 
-static register_setting_t cc1120SettingsStd[] = {
-    // Set GPIO 0 to RXFIFO_THR_PKT
-    {CC1120_REGS_IOCFG0, 0x01U},
-    // Set GPIO 1 to HighZ
-    {CC1120_REGS_IOCFG1, 0x30U},
-    // Set GPIO 2 to PKT_SYNC_RXTX
-    {CC1120_REGS_IOCFG2, 0x06U},
-    // Set GPIO 3 to TXFIFO_THR_PKT
-    {CC1120_REGS_IOCFG3, 0x03U},
-    // Set the sync word as 16 bits and allow for < 2 bit error on sync word
-    {CC1120_REGS_SYNC_CFG0, 0x09U},
-    // Set sync word qualifier value threshold similar to the one talked about for preamble in section 6.8
-    {CC1120_REGS_SYNC_CFG1, 0x08U},
-    // Set first 8 bits of the sync word to 0x55U (arbitrary value)
-    {CC1120_REGS_SYNC0, 0x55U},
-    // Set next 8 bits of the sync word to 0x57U (arbitrary value)
-    {CC1120_REGS_SYNC1, 0x57U},
-    // Set cc1120 to switch to FSTXON state after a packet is received
-    {CC1120_REGS_RFEND_CFG1, 0x1F},
-    {CC1120_REGS_DEVIATION_M, 0x3AU},
-    {CC1120_REGS_MODCFG_DEV_E, 0x0AU},
-    {CC1120_REGS_DCFILT_CFG, 0x1CU},
-    // Set the preamble as 4 bytes of 10101010
-    {CC1120_REGS_PREAMBLE_CFG1, 0x18U},
-    // enable preamble and set the error threshold
-    {CC1120_REGS_PREAMBLE_CFG0, 0x2AU},
-    {CC1120_REGS_IQIC, 0xC6U},
-    {CC1120_REGS_CHAN_BW, 0x08U},
-    {CC1120_REGS_MDMCFG0, 0x05U},
-    {CC1120_REGS_SYMBOL_RATE2, 0x73U},
-    {CC1120_REGS_AGC_REF, 0x20U},
-    {CC1120_REGS_AGC_CS_THR, 0x19U},
-    {CC1120_REGS_AGC_CFG2, 0x20U},
-    {CC1120_REGS_AGC_CFG1, 0xA9U},
-    {CC1120_REGS_AGC_CFG0, 0xCFU},
-    {CC1120_REGS_FIFO_CFG, TXRX_INTERRUPT_THRESHOLD},
-    {CC1120_REGS_FS_CFG, 0x14U},
-    {CC1120_REGS_PKT_CFG0, 0x00U},
-    {CC1120_REGS_PA_CFG0, 0x7DU},
-    {CC1120_REGS_PKT_LEN, 0x0CU}};
-
-static register_setting_t cc1120SettingsExt[] = {
-    {CC1120_REGS_EXT_IF_MIX_CFG, 0x00U}, {CC1120_REGS_EXT_FREQOFF_CFG, 0x34U}, {CC1120_REGS_EXT_FREQ2, 0x6CU},
-    {CC1120_REGS_EXT_FREQ1, 0x7AU},      {CC1120_REGS_EXT_FREQ0, 0xE1U},       {CC1120_REGS_EXT_FS_DIG1, 0x00U},
-    {CC1120_REGS_EXT_FS_DIG0, 0x5FU},    {CC1120_REGS_EXT_FS_CAL1, 0x40U},     {CC1120_REGS_EXT_FS_CAL0, 0x0EU},
-    {CC1120_REGS_EXT_FS_DIVTWO, 0x03U},  {CC1120_REGS_EXT_FS_DSM0, 0x33U},     {CC1120_REGS_EXT_FS_DVC0, 0x17U},
-    {CC1120_REGS_EXT_FS_PFD, 0x50U},     {CC1120_REGS_EXT_FS_PRE, 0x6EU},      {CC1120_REGS_EXT_FS_REG_DIV_CML, 0x14U},
-    {CC1120_REGS_EXT_FS_SPARE, 0xACU},   {CC1120_REGS_EXT_FS_VCO0, 0xB4U},     {CC1120_REGS_EXT_XOSC5, 0x0EU},
-    {CC1120_REGS_EXT_XOSC1, 0x03U},      {CC1120_REGS_EXT_TOC_CFG, 0x89U}};
-
-void initAllTxRxSemaphores(void) {
+void initAllCc1120TxRxSemaphores(void) {
   if (txSemaphore == NULL) {
     txSemaphore = xSemaphoreCreateBinaryStatic(&txSemaphoreBuffer);
     // Initialize semaphore with count of 1
@@ -106,71 +55,6 @@ void initAllTxRxSemaphores(void) {
   if (syncReceivedSemaphore == NULL) {
     syncReceivedSemaphore = xSemaphoreCreateBinaryStatic(&syncReceivedSemaphoreBuffer);
   }
-}
-
-/**
- * @brief Gets the number of bytes queued in the TX FIFO
- *
- * @param numBytes - A pointer to an 8-bit integer to store the number of bytes in
- * @return obc_error_code_t - Whether or not the registe read was successful
- */
-obc_error_code_t cc1120GetBytesInTxFifo(uint8_t *numBytes) {
-  if (numBytes == NULL) {
-    return OBC_ERR_CODE_INVALID_ARG;
-  }
-  obc_error_code_t errCode;
-  RETURN_IF_ERROR_CODE(cc1120ReadExtAddrSpi(CC1120_REGS_EXT_NUM_TXBYTES, numBytes, 1));
-  return OBC_ERR_CODE_SUCCESS;
-}
-
-/**
- * @brief Gets the state of the CC1120 from the MARCSTATE register
- *
- * @param stateNum - A pointer to an 8-bit integer to store the state in
- * @return obc_error_code_t - Whether or not the register read was successful
- */
-obc_error_code_t cc1120GetState(cc1120_state_t *stateNum) {
-  if (stateNum == NULL) {
-    return OBC_ERR_CODE_INVALID_ARG;
-  }
-  obc_error_code_t errCode;
-  RETURN_IF_ERROR_CODE(cc1120ReadExtAddrSpi(CC1120_REGS_EXT_MARCSTATE, stateNum, 1));
-  *stateNum &= 0b11111;
-  return OBC_ERR_CODE_SUCCESS;
-}
-
-/**
- * @brief Resets CC1120 & initializes transmit mode
- *
- * @return obc_error_code_t - Whether or not the setup was a success
- */
-obc_error_code_t cc1120Init(void) {
-  obc_error_code_t errCode;
-
-  RETURN_IF_ERROR_CODE(cc1120StrobeSpi(CC1120_STROBE_SRES));
-
-  // When changing which signals are sent by each gpio, the output will be unstable so interrupts should be disabled
-  // see chapter 3.4 in the datasheet for more info
-  gioDisableNotification(gioPORTB, CC1120_RX_THR_PKT_gioPORTB_PIN);
-  gioDisableNotification(gioPORTB, CC1120_TX_THR_PKT_hetPORT1_PIN);
-  gioDisableNotification(gioPORTA, CC1120_PKT_SYNC_RXTX_hetPORT1_PIN);
-
-  for (uint8_t i = 0; i < sizeof(cc1120SettingsStd) / sizeof(register_setting_t); i++) {
-    RETURN_IF_ERROR_CODE(cc1120WriteSpi(cc1120SettingsStd[i].addr, &cc1120SettingsStd[i].val, 1));
-  }
-
-  // enable interrupts again now that the gpio signals are set
-  gioEnableNotification(gioPORTB, CC1120_RX_THR_PKT_gioPORTB_PIN);
-  gioEnableNotification(gioPORTB, CC1120_TX_THR_PKT_hetPORT1_PIN);
-  gioEnableNotification(gioPORTA, CC1120_PKT_SYNC_RXTX_hetPORT1_PIN);
-  for (uint8_t i = 0; i < sizeof(cc1120SettingsExt) / sizeof(register_setting_t); i++) {
-    RETURN_IF_ERROR_CODE(cc1120WriteExtAddrSpi(cc1120SettingsExt[i].addr, &cc1120SettingsExt[i].val, 1));
-  }
-
-  initAllTxRxSemaphores();
-
-  RETURN_IF_ERROR_CODE(cc1120StrobeSpi(CC1120_STROBE_SFSTXON));
-  return OBC_ERR_CODE_SUCCESS;
 }
 
 /**
@@ -321,29 +205,13 @@ static obc_error_code_t writeFifoBlocking(uint8_t *data, uint32_t len) {
   }
   return OBC_ERR_CODE_SUCCESS;
 }
-/* RX functions */
-
-/**
- * @brief Gets the number of bytes queued in the RX FIFO
- *
- * @param numBytes - A pointer to an 8-bit integer to store the number of bytes in
- * @return obc_error_code_t - Whether or not the register read was successful
- */
-obc_error_code_t cc1120GetBytesInRxFifo(uint8_t *numBytes) {
-  if (numBytes == NULL) {
-    return OBC_ERR_CODE_INVALID_ARG;
-  }
-  obc_error_code_t errCode;
-  RETURN_IF_ERROR_CODE(cc1120ReadExtAddrSpi(CC1120_REGS_EXT_NUM_RXBYTES, numBytes, 1));
-  return OBC_ERR_CODE_SUCCESS;
-}
 
 /**
  * @brief Switches the cc1120 to RX mode to continuously receive bytes and send them to the decode task
  * @param syncWordTimeoutTicks - The amount of time to wait for the syncReceivedSemaphore to become available
  * @return obc_error_code_t
  */
-obc_error_code_t cc1120Receive(TickType_t syncWordTimeoutTicks) {
+obc_error_code_t cc1120ReceiveToDecodeTask(TickType_t syncWordTimeoutTicks) {
   obc_error_code_t errCode = OBC_ERR_CODE_SUCCESS;
   if (rxSemaphore == NULL) {
     return OBC_ERR_CODE_INVALID_STATE;
@@ -419,6 +287,80 @@ obc_error_code_t cc1120Receive(TickType_t syncWordTimeoutTicks) {
 }
 
 /**
+ * @brief Switches the cc1120 to RX mode to continuously receive bytes and send them to the decode task
+ * @param syncWordTimeoutTicks - The amount of time to wait for the syncReceivedSemaphore to become available
+ * @param recvBuf: the buffer to store the received bytes
+ * @param recvBufLen: length of recvBuf
+ * @return obc_error_code_t
+ */
+obc_error_code_t cc1120Receive(uint8_t *recvBuf, uint16_t recvBufLen, TickType_t syncWordTimeoutTicks) {
+  if (rxSemaphore == NULL) {
+    return OBC_ERR_CODE_INVALID_STATE;
+  }
+  if (recvBuf == NULL) {
+    return OBC_ERR_CODE_INVALID_ARG;
+  }
+
+  obc_error_code_t errCode = OBC_ERR_CODE_SUCCESS;
+
+  // poll the semaphore to clear whatever value it has (do not block and wait on it)
+  xSemaphoreTake(syncReceivedSemaphore, (TickType_t)0);
+
+  // When changing which signals are sent by each gpio, the output will be unstable so interrupts should be disabled
+  // see chapter 3.4 in the datasheet for more info
+  gioDisableNotification(gioPORTA, CC1120_PKT_SYNC_RXTX_hetPORT1_PIN);
+
+  // switch gpio 2 to be a SYNC_EVENT signal instead of CC1120_PKT_SYNC_RXTX_PIN
+  uint8_t spiTransferData = SYNC_EVENT_SIGNAL_NUM;
+  RETURN_IF_ERROR_CODE(cc1120WriteSpi(CC1120_REGS_IOCFG2, &spiTransferData, 1));
+
+  // enable interrupts again now that the gpio signals are set
+  gioEnableNotification(gioPORTA, (uint32_t)CC1120_SYNC_EVENT_PIN);
+
+  // Temporarily set packet size to infinite
+  spiTransferData = INFINITE_PACKET_LENGTH_MODE;
+  RETURN_IF_ERROR_CODE(cc1120WriteSpi(CC1120_REGS_PKT_CFG0, &spiTransferData, 1));
+
+  // Switch cc1120 to receive mode
+  RETURN_IF_ERROR_CODE(cc1120StrobeSpi(CC1120_STROBE_SRX));
+
+  // wait to receive sync word before continuing
+  if (xSemaphoreTake(syncReceivedSemaphore, syncWordTimeoutTicks) != pdPASS) {
+    LOG_ERROR_CODE(OBC_ERR_CODE_SEMAPHORE_TIMEOUT);
+    return OBC_ERR_CODE_SEMAPHORE_TIMEOUT;
+  }
+  // See chapters 8.1, 8.4, 8.5
+  // If we do not stop receiving data, continue looping until COMMS_MAX_UPLINK_BYTES rounded up to the nearest multiple
+  // of TXRX_INTERRUPT_THRESHOLD bytes are received
+  uint8_t rxFifoReadCycles;  // number of times we receive TXRX_INTERRUPT_THRESHOLD bytes and read them out
+  for (rxFifoReadCycles = 0; rxFifoReadCycles < (recvBufLen + TXRX_INTERRUPT_THRESHOLD - 1) / TXRX_INTERRUPT_THRESHOLD;
+       ++rxFifoReadCycles) {
+    // wait until we have not received more than TXRX_INTERRUPT_THRESHOLD bytes for more than rxTimeout
+    // before exiting this loop since that means we are no longer transmitting
+    if (xSemaphoreTake(rxSemaphore, RX_SEMAPHORE_TIMEOUT) != pdPASS) {
+      break;
+    }
+    RETURN_IF_ERROR_CODE(cc1120ReadFifo(recvBuf, TXRX_INTERRUPT_THRESHOLD));
+  }
+
+  uint8_t numBytesInRxFifo = 0;
+
+  // check the number of bytes remaining in the RX FIFO
+  RETURN_IF_ERROR_CODE(cc1120GetBytesInRxFifo(&numBytesInRxFifo));
+
+  if (numBytesInRxFifo > recvBufLen - (rxFifoReadCycles * TXRX_INTERRUPT_THRESHOLD)) {
+    return OBC_ERR_CODE_BUFF_TOO_SMALL;
+  }
+
+  if (numBytesInRxFifo != 0) {
+    // if there are still bytes in the RX FIFO, read them out
+    RETURN_IF_ERROR_CODE(cc1120ReadFifo(recvBuf, numBytesInRxFifo));
+  }
+
+  return OBC_ERR_CODE_SUCCESS;
+}
+
+/**
  * @brief block until the tx fifo is empty without decrementing the semaphore
  *
  * @return obc_error_code_t - whether the tx fifo empty semaphore became available without timing out or not
@@ -478,5 +420,3 @@ void syncEventCallback(void) {
   // been unblocked
   portYIELD_FROM_ISR(xHigherPriorityTaskAwoken);
 }
-
-SemaphoreHandle_t getCC1120RxSemaphoreHandle(void) { return rxSemaphore; }
